requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
cfgv==3.1.0
click==7.1.1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
identify==1.4.13
mccabe==0.6.1
nodeenv==1.3.5
pathspec==0.7.0
pre-commit==2.2.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
regex==2020.2.20
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
virtualenv==20.0.15
